Update productForSuggestionsSchema to enforce integer values for price, ensuring it is non-negative

+++ backend/schemas/productforsuggestions.schema.ts
@@ -3,7 +3,7 @@ import z from "zod";
 export const productForSuggestionsSchema = z.object({
     name: z.string().min(1, "Nome é obrigatório"),
     description: z.string().optional(),
-    price: z.number().min(0, "Preço deve ser zero ou maior"),
+    price: z.number().int().min(0, "Preço deve ser zero ou maior"),
     stock: z.number().int().min(0, "Estoque deve ser zero ou maior"),
 });
